debate_paradise/requirements.txt
Django==4.2
whitenoise>=6.6,<7
djangorestframework==3.14
djangorestframework-simplejwt>=5.3.1
django-cors-headers==4.0
PyMySQL==1.1.1
django-redis==5.3
celery==5.3
setuptools>=69
bcrypt>=4.1.3
